requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
schedule==1.2.1
jdatetime==4.1.1
openpyxl==3.0.9
pytz==2023.3
requests